Test to try and hide more of the SDK details from the rest of the CPP app.
Introduce RoomStream as a way of subscribing to room updates from a connection. The idea is to hide dispatcher as this will likely become very complicated as the bindings grow, similarly roommodel no longer needs to know the underlying bindings for RoomList it just gets a stream of diffs via signals which is similar to what a quotient room gives us now.